[MWSE] Phase E.B.3 batch 18: reconcile NITriShape, shim it
NITriShape reconciliation -- bring SharedSE form up to MWSE-original
surface so the shim builds clean. The SharedSE form had a different
ctor/create signature path than MWSE-original; combine them as
overloads rather than choosing one, gating the SharedSE-only ones to
CSSE so MWSE doesn't expect an impl that doesn't exist there.

Additions to SharedSE/NITriShape.h (all from MWSE-original):
- namespace TriShapeFlags { enum { SoftwareSkinningFlag = 0x200 }; }
- TriShape(TriBasedGeometryData* data) ctor (single-arg)
- void linkObject(Stream* stream)
- static Pointer<TriShape> create(unsigned short vertexCount,
                                  bool hasNormals, bool hasColors,
                                  unsigned short textureCoordSets,
                                  unsigned short triangleCount)
- MWSE_SOL_CUSTOMIZED_PUSHER_DECLARE_NI(NI::TriShape) under SE_USE_LUA

Gated to CSSE-only (!SE_IS_MWSE):
- The many-arg ctor + dtor + matching create overload. SharedSE has no
  NITriShape.cpp -- these are aspirational decls. MWSE doesn't impl
  them either, so leaving them visible would trigger LNK2001 on the
  dtor via sol's new_usertype<TriShape> in NITriShapeLua.cpp (same
  pattern as Lines/Pick/TexturingProperty in earlier batches).

Field types in getVertices/getNormals: unify to nonstd::span<Vector3>.
NI::Vector3 == TES3::Vector3 in MWSE via typedef bridge, so this
matches MWSE-private NITriShape.cpp's TES3::Vector3 return.

Shimmed:
- NITriShape.h

Build verified clean (MWSE Release|x86, 0 errors).

Phase E.B.3 status: 67 of 70 NI headers shimmed. Remaining unshimmed
are NIDefines.h (intentionally MWSE-private; carries VirtualTableAddress
enum) plus any I missed in the survey.
